Navigation is delegated to robot position movement.
Navigation is a timer that will move depending on its moveToX and moveToY attributes.
Potential changes: Navigation really is just a motor movement rather than an actual navigation controller. Could be renamed PositionMovement and actual logic of navigation is done in a environment reaction class, which changes the moveToX and moveToY attributes as needed.