CollectionSystem controls and executes block collection.
Our design has not been defined so nothing is set in stone here, but there is a constraint of the number of motors available depending on number of NXT bricks. Touch sensors and other sensors could be used and turned off as needed. Once this thread is called reaction to environment should be paused.